+++ src/Oro/Bundle/OrderBundle/Tests/Functional/Controller/OrderControllerTest.php
@@ -104,7 +104,7 @@ class OrderControllerTest extends WebTestCase
 
     protected function setUp()
     {
-        $this->initClient([], array_merge($this->generateBasicAuthHeader(), ['HTTP_X-CSRF-Header' => 1]));
+        $this->initClient([], $this->generateBasicAuthHeader());
         $this->client->useHashNavigation(true);
 
         $this->loadFixtures(